+++ hooks/pre-commit.sh
@@ -47,6 +47,18 @@ else
     echo ""
 fi
 
+echo "$(tput setaf 3)* Does not cntain .only in tests?$(tput sgr 0)"
+grep -ri --exclude-dir=with-typescript "\.only" ./test
+dotOnly=$?
+
+if [ $dotOnly -eq 1 ]
+then
+   echo "$(tput setaf 2)* Yes$(tput sgr 0)"
+else
+   echo "$(tput setaf 1)* No$(tput sgr 0)"
+   echo ""
+fi
+
 if [ $no_of_files_to_stash -ne 0 ]
 then
    echo "$(tput setaf 3)* Undoing stashing$(tput sgr 0)"
@@ -58,19 +70,26 @@ then
    git stash drop >/dev/null 2>/dev/null
 fi
 
-if [ $compiles -eq 0 ] && [ $formatted -eq 0 ]
+if [ $compiles -eq 0 ] && [ $formatted -eq 0 ] && [ $dotOnly -eq 1 ]
 then
    echo "$(tput setaf 2)... done. Proceeding with commit.$(tput sgr 0)"
    echo ""
-elif [ $compiles -eq 0 ]
+elif [ $compiles -ne 0 ]
+then
+   echo "$(tput setaf 1)... done.$(tput sgr 0)"
+   echo "$(tput setaf 1)CANCELLING commit due to COMPILE ERROR.$(tput sgr 0)"
+   echo ""
+   exit 2
+elif [ $formatted -ne 0 ]
 then
    echo "$(tput setaf 1)... done.$(tput sgr 0)"
    echo "$(tput setaf 1)CANCELLING commit due to NON-FORMATTED CODE.$(tput sgr 0)"
    echo ""
    exit 1
-else
+elif [ $dotOnly -ne 1 ]
+then
    echo "$(tput setaf 1)... done.$(tput sgr 0)"
-   echo "$(tput setaf 1)CANCELLING commit due to COMPILE ERROR.$(tput sgr 0)"
+   echo "$(tput setaf 1)CANCELLING commit due to .only in test files $(tput sgr 0)"
    echo ""
    exit 2
 fi