+++ lib/ts/recipe/thirdparty/providers/apple.ts
@@ -141,7 +141,6 @@ export default function Apple(config: TypeThirdPartyProviderAppleConfig): TypePr
             let supertokens = SuperTokens.getInstanceOrThrowError();
             return (
                 supertokens.appInfo.apiDomain.getAsStringDangerous() +
-                supertokens.appInfo.apiGatewayPath.getAsStringDangerous() +
                 supertokens.appInfo.apiBasePath.getAsStringDangerous() +
                 APPLE_REDIRECT_HANDLER
             );