+++ CHANGELOG.md
@@ -13,6 +13,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Adds passwordless tests
 -   passwordless test fixs
 
+## [8.3.0] - 2021-12-08
+
+### Added
+
+-   The ability to enable JWT creation with session management, this allows easier integration with services that require JWT based authentication: https://github.com/supertokens/supertokens-core/issues/250
+
+## [8.2.1] - 2021-12-07
+
+### Fixes
+
+-   Removes use of apiGatewayPath from apple's redirect URI since that is already there in the apiBasePath
+
 ## [8.2.0] - 2021-11-15
 
 ### Added